+++ Unity/Assets/Scripts/HotfixView/Client/MengJing/UI/UIItemTip/UIItemTipComponentSystem.cs
@@ -35,19 +35,19 @@ namespace ET.Client
             if (item != null)
             {
                 ItemConfig itemConfig = ItemConfigCategory.Instance.Get(item.ConfigId);
-                if (itemConfig.ItemType == (int)ItemType.Consume)
+                if (itemConfig.ItemType == ItemType.Consume)
                 {
                     self.UIItemTip_ConsumeComponent.GameObject.SetActive(true);
                     self.UIItemTip_ConsumeComponent.UpdateInfo(uiItemTipData).Coroutine();
                 }
 
-                if (itemConfig.ItemType == (int)ItemType.Material)
+                if (itemConfig.ItemType == ItemType.Material)
                 {
                     self.UIItemTip_MaterialComponent.GameObject.SetActive(true);
                     self.UIItemTip_MaterialComponent.UpdateInfo(uiItemTipData).Coroutine();
                 }
 
-                if (itemConfig.ItemType == (int)ItemType.Equipment)
+                if (itemConfig.ItemType == ItemType.Equipment)
                 {
                     self.UIItemTip_EquipmentComponent.GameObject.SetActive(true);
                     self.UIItemTip_EquipmentComponent.UpdateInfo(uiItemTipData).Coroutine();
